LayoutTests/fast/element-targeting/target-iframe-above-nav-bar.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<meta name="viewport" content="width=device-width, initial-scale=1">
<head>
<script src="../../resources/ui-helper.js"></script>
<script src="../../resources/js-test.js"></script>
<style>
body, html {
    margin: 0;
    font-family: system-ui;
}

.nav {
    position: sticky;
    top: 0;
    left: 0;
    width: 100%;
    height: 50px;
    background: gray;
    color: white;
    text-align: center;
    font-size: 20px;
    line-height: 50px;
}

.content {
    margin-top: 100px;
}

.frame-wrapper {
    width: 100%;
    height: 150px;
}

iframe {
    width: 100%;
    height: 100%;
}
</style>
</head>
<body>
<header>
    <div class="frame-wrapper">
        <iframe frameborder="0" srcdoc="<head><style>body, html { background: tomato; width: 100%; height: 100%; margin: 0; }</style></head><body></body>"></iframe>
    </div>
    <div class="nav" role="navigation">Navigation Bar</div>
</header>
<p class="content">This test requires WebKitTestRunner.</p>
<pre id="console"></pre>
<script>
jsTestIsAsync = true;
addEventListener("load", async () => {
    targetSelector = await UIHelper.adjustVisibilityForFrontmostTarget(50, 100);
    shouldBeEqualToString("targetSelector", "DIV.frame-wrapper");
    finishJSTest();
});
</script>
</body>
</html>
