LayoutTests/fast/element-targeting/target-iframe-above-nav-bar-expected.txt


Navigation Bar
This test requires WebKitTestRunner.

PASS targetSelector is "DIV.frame-wrapper"
PASS successfullyParsed is true

TEST COMPLETE
